+++ Security/Firewall/WsseListener.php
@@ -2,8 +2,9 @@
 namespace Oka\ApiBundle\Security\Firewall;
 
 use Oka\ApiBundle\Security\Authentication\Token\WsseUserToken;
+use Oka\ApiBundle\Util\ErrorResponseFactory;
 use Oka\ApiBundle\Util\LoggerHelper;
-use Oka\ApiBundle\Util\ResponseHelper;
+use Oka\ApiBundle\Util\RequestHelper;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
@@ -27,16 +28,10 @@ class WsseListener extends LoggerHelper implements ListenerInterface
 	 */
 	protected $authenticationManager;
 	
-	/**
-	 * @var ResponseHelper $responseHelper
-	 */
-	protected $responseHelper;
-	
-	public function __construct(TokenStorageInterface $tokenStorage, AuthenticationManagerInterface $authenticationManager, ResponseHelper $responseHelper)
+	public function __construct(TokenStorageInterface $tokenStorage, AuthenticationManagerInterface $authenticationManager)
 	{
 		$this->tokenStorage = $tokenStorage;
 		$this->authenticationManager = $authenticationManager;
-		$this->responseHelper = $responseHelper;
 	}
 	
 	public function handle(GetResponseEvent $event)
@@ -58,7 +53,6 @@ class WsseListener extends LoggerHelper implements ListenerInterface
 			try {
 				$authToken = $this->authenticationManager->authenticate($token);
 				$this->tokenStorage->setToken($authToken);
-				
 				return;
 			} catch (\Exception $e) {
 				$failedMessage = $e->getMessage();
@@ -72,7 +66,7 @@ class WsseListener extends LoggerHelper implements ListenerInterface
 		}
 		
 		// Deny authentication with a '403 Forbidden' HTTP response
-		$content =  ResponseHelper::createError(Response::HTTP_FORBIDDEN, $failedMessage);
-		$event->setResponse($this->responseHelper->getAcceptableResponse($request, $content, Response::HTTP_FORBIDDEN, ['X-Secure-With' => 'WSSE'], 'json'));
+		$format = RequestHelper::getFirstAcceptableFormat($request) ?: 'json';
+		$event->setResponse(ErrorResponseFactory::create($failedMessage, Response::HTTP_FORBIDDEN, null, [], Response::HTTP_FORBIDDEN, ['X-Secure-With' => 'WSSE'], $format));
 	}
 }