+++ Util/RequestHelper.php
@@ -3,6 +3,7 @@ namespace Oka\ApiBundle\Util;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Validator\Constraint;
+use Symfony\Component\Validator\ConstraintViolationInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
@@ -17,15 +18,9 @@ class RequestHelper
 	 */
 	protected $validator;
 	
-// 	/**
-// 	 * @var Translator $translator
-// 	 */
-// 	protected $translator;
-	
-	public function __construct(ValidatorInterface $validator)//, Translator $translator)
+	public function __construct(ValidatorInterface $validator)
 	{
 		$this->validator = $validator;
-// 		$this->translator = $translator;
 	}
 	
 	/**
@@ -35,41 +30,19 @@ class RequestHelper
 	 * @param Constraint $constraints
 	 * @param array $groups
 	 * @return string[]
+	 * 
+	 * @deprecated
 	 */
 	public function isValid($value, Constraint $constraints, array $groups = null)
 	{
 		$errors = [];
 		
+		/** @var ConstraintViolationInterface $error */
 		foreach ($this->validator->validate($value, $constraints, $groups) as $error) {
 			$errors[preg_replace('#\[(.+)\]#', '$1', $error->getPropertyPath())] = $error->getMessage();
-// 			$errors[preg_replace('#\[(.+)\]#', '$1', $error->getPropertyPath())] = $this->translator->trans($error->getMessage());
 		}
 		
 		return $errors;
-	}
-	
-	/**
-	 * Get first acceptable response format
-	 * 
-	 * @param Request $request
-	 * @return string|NULL
-	 */
-	public static function getFirstAcceptableFormat(Request $request)
-	{
-		$acceptableContentTypes = $request->getAcceptableContentTypes();
-		
-		return $request->getFormat(empty($acceptableContentTypes) ? 'text/html' : $acceptableContentTypes[0]);
-	}
-	
-	/**
-	 * Sanitize request query
-	 * 
-	 * @param string $query
-	 * @return string
-	 */
-	public static function sanitizeQueryString($query)
-	{
-		return trim(rawurldecode($query));
 	}
 	
 	/**
@@ -95,18 +68,14 @@ class RequestHelper
 	}
 	
 	/**
-	 * Parse request query
+	 * Sanitize request query
 	 * 
-	 * @param Request $request
-	 * @param string $key
-	 * @param string $delimiter
-	 * @param mixed $defaultValue
-	 * @return mixed
-	 * @deprecated
+	 * @param string $query
+	 * @return string
 	 */
-	public static function parseQueryString(Request $request, $key, $delimiter = null, $defaultValue = null)
+	public static function sanitizeQueryString($query)
 	{
-		return self::parseQueryStringToArray($request, $key, $delimiter, $defaultValue);
+		return trim(rawurldecode($query));
 	}
 	
 	/**
@@ -120,12 +89,23 @@ class RequestHelper
 		switch ($request->getContentType()) {
 			case 'json':
 				return json_decode($request->getContent(), true);
-				
 			case 'form':
 				return $request->request->all();
-				
 			default;
 				return [];
 		}
+	}
+	
+	/**
+	 * Get first acceptable response format
+	 * 
+	 * @param Request $request
+	 * @return string|NULL
+	 */
+	public static function getFirstAcceptableFormat(Request $request)
+	{
+		$acceptableContentTypes = $request->getAcceptableContentTypes();
+		
+		return $request->getFormat(empty($acceptableContentTypes) ? 'text/html' : $acceptableContentTypes[0]);
 	}
 }