+++ Util/ResponseHelper.php
@@ -9,6 +9,7 @@ use Symfony\Component\HttpFoundation\Response;
  * 
  * @author cedrick
  * 
+ * @deprecated  Use instead `ErrorResponseBuilder` class or ErrorResponseFactory class.
  */
 class ResponseHelper
 {
@@ -16,7 +17,6 @@ class ResponseHelper
 	const SERVER_ERROR = 'Une erreur est survenue pendant le traitement de la requête.';
 	const UNEXPECTED_ERROR = 'Une erreur inattendu est survenue le traitement de la requête.';
 	const INVALID_CREDENTIALS = 'Le nom d\'utilisateur ou le mot de passe est incorrect.';
-	
 	const ERROR_EXTRAS = 'extras';
 	const ERROR_COLLECTION = 'errors';
 	const ERROR_PROPERTIES = 'properties';
@@ -41,9 +41,13 @@ class ResponseHelper
 	 * @param string $format
 	 * @param boolean $isError
 	 * @return \Symfony\Component\HttpFoundation\Response
+	 * 
+	 * @deprecated  Use instead `ErrorResponseBuilder` class or ErrorResponseFactory class. 
 	 */
 	public function getAcceptableResponse(Request $request, $content, $statusCode, $headers = [], $format = null, $isError = false)
 	{
+		@trigger_error('The '.__METHOD__.' method is deprecated since version 1.0.1 and will be removed in 2. Use instead `ErrorResponseBuilder` class or ErrorResponseFactory class.', E_USER_DEPRECATED);
+		
 		switch ($format ?: RequestHelper::getFirstAcceptableFormat($request)) {
 			case 'html':
 				if ($isError === true && isset($content['error'])) {
@@ -76,8 +80,13 @@ class ResponseHelper
 	 * @param array $errors
 	 * @param array $extras
 	 * @return string[]|string[][]
+	 * 
+	 * @deprecated since version 1.0.1, to be removed in 2. Use instead `ErrorResponseBuilder` class or ErrorResponseFactory class.
 	 */
-	public static function createError($code, $message, array $properties = [], array $extras = [], array $errors = []) {		
+	public static function createError($code, $message, array $properties = [], array $extras = [], array $errors = [])
+	{
+		@trigger_error('The '.__METHOD__.' method is deprecated since version 1.0.1 and will be removed in 2. Use instead `ErrorResponseBuilder` class or ErrorResponseFactory class.', E_USER_DEPRECATED);
+		
 		$error = ['error' => ['code' => $code, 'message' => $message]];
 		
 		if (!empty($errors)) {